+++ crypto/tls/config/config_test.go
@@ -1,9 +1,9 @@
-package tls_test
+package config_test
 
 import (
 	"testing"
 
-	"github.com/alexfalkowski/go-service/v2/crypto/tls"
+	tls "github.com/alexfalkowski/go-service/v2/crypto/tls/config"
 	"github.com/alexfalkowski/go-service/v2/internal/test"
 	"github.com/stretchr/testify/require"
 )
