+++ net/http/transport.go
@@ -1,10 +1,10 @@
 package http
 
 import (
-	"crypto/tls"
 	"net"
 	"net/http"
 
+	"github.com/alexfalkowski/go-service/v2/crypto/tls"
 	"github.com/alexfalkowski/go-service/v2/time"
 )
