+++ transport/http/client_test.go
@@ -3,7 +3,7 @@ package http_test
 import (
 	"testing"
 
-	"github.com/alexfalkowski/go-service/v2/crypto/tls"
+	tls "github.com/alexfalkowski/go-service/v2/crypto/tls/config"
 	"github.com/alexfalkowski/go-service/v2/internal/test"
 	"github.com/alexfalkowski/go-service/v2/transport/http"
 	"github.com/stretchr/testify/require"
