+++ transport/http/doc.go
@@ -41,8 +41,10 @@
 // # Registration gotcha (TLS filesystem)
 //
 // This package uses package-level registration to inject filesystem access used when constructing TLS
-// configuration. The registered filesystem is used to resolve certificate/key "source strings" (for example
-// `file:/path/to/cert` or `env:VAR`) when materializing `*tls.Config`.
+// configuration. The registered filesystem is used by
+// `crypto/tls/config.NewConfig` to resolve certificate/key "source strings"
+// (for example `file:/path/to/cert` or `env:VAR`) when materializing the
+// runtime `*crypto/tls.Config`.
 //
 // If you enable TLS and do not call `Register` before constructing HTTP servers or clients, TLS configuration
 // may fail to load key material.
