+++ transport/http/register.go
@@ -9,8 +9,9 @@ var fs *os.FS
 // Register injects the filesystem dependency used by this package.
 //
 // The registered filesystem is consulted when constructing TLS configuration for both clients and servers.
-// It is used to resolve certificate/key "source strings" (for example `file:/path/to/cert` or `env:VAR`) via
-// the `os.FS.ReadSource` helper when materializing `*tls.Config`.
+// It is used by `crypto/tls/config.NewConfig` to resolve certificate/key
+// "source strings" (for example `file:/path/to/cert` or `env:VAR`) via the
+// `os.FS.ReadSource` helper when materializing a runtime `*crypto/tls.Config`.
 //
 // Registration is required because the filesystem is stored in a package-level variable. If TLS is enabled,
 // call Register during application startup (composition root) before constructing any HTTP clients or servers;
